Extract registryserver code to a package
It is hard to write client code for config-resolver when the endpoint
code (like query parameter names) is in `main.go`. Hence, extract the
code to a package.
